requirements.txt
aerich==0.7.2
aio-pika==9.4.1
aiomysql==0.2.0
aioredis==2.0.1
aiormq==6.8.0
aiosqlite==0.17.0
alipay-sdk-python==3.7.78
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncmy==0.2.9
atlastk==0.13.3
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
captcha==0.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dictdiffer==0.9.0
ecdsa==0.18.0
exceptiongroup==1.2.0
Faker==24.11.0
fastapi==0.110.1
fastapi-pagination==0.12.23
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
multidict==6.0.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pamqp==3.3.0
pandas==2.2.2
passlib==1.7.4
pillow==10.3.0
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pypika-tortoise==0.1.6
PySocks==1.7.1
python-alipay-sdk==3.3.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
tomlkit==0.12.4
tortoise==0.1.1
tortoise-orm==0.20.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.21.1
wsproto==1.2.0
yarl==1.9.4
